Retry LSP tests in case of error (#5342)
### Description

This PR changes how the language server unit tests execute, retrying
them up to three times if they fail. Ultimately, this is what we're
doing by hand, anyway, so we might as well automate it.

The proper solution to this problem is to ensure that the language
server tests don't spuriously fail, but this will accelerate our
development process in the meantime.

### How has this been tested?

By running CI normally.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.