Allow using project files together with --library (#5297)
### Changes
- To enable smoothly working with multiple projects inside a single
repository, Dafny now allows using a Dafny project file as an argument
to `--library`. This does the following:
1. Call `dafny build -t:lib` using the project file to get a doo file.
In the future we plan to let this be a NOOP if the right build file is
already there.
1. Replace the usage of the project file in the origin call to `dafny`,
with this doo file.
- Extract code from `DafnyFile.CreateAndValidate` into methods
- When parsing a file, correctly pass parse options to included files.
This does not have any particular affect because only doo files are
passed with different options, and those do not use includes. However,
doing this does seem more correct and it was useful previously, so I'll
keep it in.
- Move `ReadsClausesOnMethods` option code
- Do not let project files include themselves through something like
`includes = ["**/*"]`

### How has this been tested?
- Add the CLI test `projectAsLibrary.dfy`

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>