Revert "Retry LSP tests in case of error (#5342)" (#5361)
This reverts commit 44ca1b2a165c69a32f39714e95cdd07002da2ebb.

### Description
The recently merged PR https://github.com/dafny-lang/dafny/pull/5342
argues that we should automatically retry failing IDE tests, because
that's what we manually do anyways. However, we only do that in some
cases. Specifically, when any of us make changes that may create
unstable tests, such as changes in the IDE, then failing IDE tests
should not be retried, manually or automatically. PR5342 makes it so
that such a retry happens automatically, thereby making it much easier
to introduce non-deterministic bugs either in IDE code or tests. That's
why this PR is reverting that one.

IMO there is little difference between a test that fails 100% of the
time and one that fails <100% of the time. Retrying a failing test that
fails <100% of the time is similar to saying "I believe my PR is not at
fault for this test failing, please ignore it". A similar situation
occurs when a PR gets merged that causes nightly to fail. However then
we don't have a way for new PRs to say "please ignore nightly failing",
and we explicitly block all PR until nightly is fixed again. A similar
approach to unstable tests would be to not to retry failing builds until
the instability is resolved, but instead to wait until an instability
fix PR is merged.

### How has this been tested?
It's a revert.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>