Unsafe dependencies (#5394)
### Description
Introduce a new option `--unsafe-dependencies` which is always on for
`dafny server` and `dafny resolve`, optionally enabled for `dafny
verify`, and off for all other commands. This option only affects
compilation when a dependency (--library) is specified using a Dafny
project file. In case the option is off, that dependency project is used
to build a library file, which is then used as the dependency. In case
the option is on, the source files from the dependency project are
included but not verified or translated.

The goal of this PR is to to enable practical development of Dafny
projects where multiple projects that depend on each other live inside a
single repository, while not affecting the CI of Dafny projects in any
way.

### How has this been tested?
- Updated the test projectAsLibrary

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>