Feat: Rust operators (#5390)
UPDATE: The commit
https://github.com/dafny-lang/dafny/pull/5390/commits/8661eb02a51238c8aad15306ef0e1fd15161a231
fixed the previous nightly issue that multi-compiler tests on Windows
were taking 15x longer.
UPDATE: That works only for the ResolvedDesugaredExecutableDafny
compiler. Enabling the Dafny-to-Rust compiler again triggers that 15x
slowdown. **Do not merge yet**

### Description
This PR implements most immutable types in Dafny (Sequence, Set,
Multiset, Map) and implements most if not all operators on them. It also
sets an architecture that can now be relied on.
List of features in this PR:
- More comprehensive Rust AST, precedence and associativity to render
parentheses
- The Rust compiler no longer crash when not in /runAllTests, but emits
invalid code. It makes it possible to see which features are missing by
searching for `<i>` in the generated code.
- Made verification of GenExpr less brittle
- Previous PR review comments
- Ability to use shadowed identifiers to avoid numbers in names. Rust
has the same rules as Dafny for shadowing, and I will continue to
monitor cases where semantics might differ.
- `make pr` now prepares a codebase for a (regenerates everything and
reformats everything in the correct order).

### How has this been tested?
I added an integration test for compiling Dafny to Rust, and I added
unit tests for the runtime in Rust
*All tests for each compiler now have a `.rs.check` if they fail.*.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>