Implement GetAssertedExpr for  DecreasesBoundedBelow
This introduces more duplication than I’d like, but it’ll be a bit of
work to remove it. Ideally, the Boogie code would be generated by
translating the Dafny code that’s now present, but that would likely
change the generated Boogie code in semantically equivalent ways that
nonetheless might perturb verification. It’s also not clear that it’s
feasible to make the variable declarations that store the old value of
loop decreases expressions come from translated Dafny code.