Implement `decreases to` expressions (#5367)
### Description

Implement the `decreases to` and `nonincreases to` expression forms.
This allows you to state that the left-hand sequence of expressions
decreases (or at least doesn't increase) to the right-hand sequence.

One caveat: for now, `decreases to` is allowed only inside parentheses,
to avoid parsing ambiguities. This is a strict subset of what we'd
ideally like to support. I think it should be possible to allow it to
appear elsewhere as part of a separate, backward-compatible PR.

Fixes #5252

### How has this been tested?

Added `dafny0/DecreasesTo{1,2,3,4,5}.dfy`

By submitting this pull request, I confirm that my contribution is made
under the terms of the MIT license.