feat: python-module-name (#5461)
### Description
<!-- Is this a user-visible change? Remember to update RELEASE_NOTES.md
-->
<!-- Is this a bug fix for an issue visible in the latest release?
Mention this in the PR details and ensure a patch release is considered
-->

#### 1

Add a `--python-module-name` option to the (new) Dafny CLI.
This option takes in a string that will be prefixed to translated module
imports.

ex. If a module has a generated Python import statement `import
MyModule`,
then translating with `--python-module-name=my_python_module`
will result in an import statement `import my_python_module.MyModule as
MyModule`.

I intend that Smithy-Dafny will use this with
`--python-module-name=[module_name].internaldafny.generated`.

#### 2

Do not generate `import module_` for `System_.py` in the DafnyRuntime
Python.
Before this change `System_.py` in the DafnyRuntimePython generates this
import. This removes that import.
Two problems with this import:
1. (Correctness problem, i.e. "why this change is necessary") The
DafnyRuntimePython does not include a `module_.py`. Having `import
module_` prevents the DafnyRuntimeModule from being used on its own.
2. (My problem, i.e. "why I'm making this change now") `import module_`
is an invalid import statement with module mode. `module_` will be
located at `[python_module_name].module_`, and isn't able to be imported
without qualifiers.

An alternative to removing the import is to generate a `module_.py` in
the DafnyRuntimePython. I don't know why DafnyRuntimePython doesn't have
this file, so this was easier for me. I don't have an opinion on which
approach happens; I'll defer to the Dafny team on the approach.

#### 3

Python: Do not generate import statements for empty modules.
"Empty modules" might come from prefixes of nested modules.
This is a follow-up to
https://github.com/dafny-lang/dafny/commit/4ce2f7e99e8c1ba6cdaa104e9943036587f703ce,
which stopped generating empty modules.

### How has this been tested?
<!-- Tests can be added to
`Source/IntegrationTests/TestFiles/LitTests/LitTest/` or to
`Source/*.Test/…` and run with `dotnet test` -->

Some tests in the directory above:

* singlemodule (validating compiling and running with 1 module)
* multimodule (validating compiling and running with 2 modules with a
dependency relationship)
* nestedmodule (validating depending on a (Python) module with a nested
(Dafny) module name)

I've tested this with Smithy-Dafny-Python and the (WIP) Python MPL, and
this behavior works as I'd expect.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Shubham Chaturvedi <anotherosscontributor@gmail.com>
Co-authored-by: Robin Salkeld <salkeldr@amazon.com>