Optimize compilation of functional-looking assignment RHSs (#5589)
### Description

Previously, the single-pass compiler would use `TrExprOpt` to turn
nested expressions into target-language statements, ending in a
`return`. This PR generalizes that mechanism so that the ending action
is parameterized. In particular, the PR now applies `TrExprOpt` to
single-LHS assignment statements. This allows the RHS of such assignment
statements to be deeply nested expressions, without running into, for
example, the limits of `javac` to deal with nested lambda expressions.

Fixes #3868

### Implementation

This PR required a change to the target-code strategy of `match`
constructs. Previously, `case`s turned into a sequence of `if`
statements (note, not a cascading `if`). If one of those `if` statements
determined that the `case` applied, then a boolean variable,
`unmatched`, was set to `false`. Each of these `if` statements was thus
also guarded by the condition `unmatched`.

This previous strategy caused a problem when the body of the cases is an
assignment. Since every assignment was guarded by an `if`, target
languages with definite-assignment rules weren't able to determine that
the variable was always assigned.

A more straightforward strategy is to jump (using a `break` or `goto`)
to the end of the target code for the `match`, rather than setting the
`unmatched` variable. This lets the target language easily see that the
variable has been assigned on all paths. It also means that the
`unmatched` variable and its tests are not needed at all. Furthermore,
the last `case` does not need an enclosing `if` structure at all, since
if the prior `case`s don't apply, then the last one must apply.

This PR improves the strategy to the one with jumps.

### Work along the way

fix: Use Downcast for let RHS and for function body

Fixes #5597

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>