Chore: better compile-time elephant operator (#5614)
### Description
This is not a user-visible change, only a refactoring that enables all
compilers to perform a rewriting optimization.
Before,
```
var o :- B;
```
was rewritten into
```
var o;
var tmp := B;
if tmp.IsFailure() {
  return tmp.PropagateFailure();
}
o := tmp.Extract();
```
The problem with this layout is that "o" is declared long before it's
assigned. It's even declared when it's not assigned ! Because compilers
often assign a placebo value to "o", this value is useless if `B` leads
to a failure.
Moreover, in the case of the Rust compiler where we want to use Dafny
names for variables, the above caused an issue in that if `o` was
declared before and used in `B`, the generated code would capture the
new definition of `o`, leading at best into a type-checking issue, at
worst into a run-time exception.

For these two reasons, this PR rewrites the above into:
```
var tmp := B;
if tmp.IsFailure() {
  return tmp.PropagateFailure();
}
var o := tmp.Extract();
```
at compile-time. It does so by following the rules of the `:-` operator,
namely only the first variable might be late-declared that way.

The PR ensures that the act of not declaring one of the variables is
certainly followed by declaring it before the call to `.Extract()`. The
code first detect if such a rewriting is possible at the `VarDeclStmt`.
If so, it stores the index on the resolved statements so that the
translation of the inner `UpdateStmt` knows to declare `o` when it
assigns it.

Previous attempts at modifying the list of resolved statements failed
several times. In this approach, we ensure that we don't modify the
resolver or the resolved code, only optimize the code during a
compilation phase common to all back-ends.

### How has this been tested?
All CI tests should pass. I verified that all the new branches were
taken in the test `exceptions/Exceptions1.dfy` with the debugger. Later,
https://github.com/dafny-lang/dafny/pull/5613 will add two tests that
would have failed without this feature.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Siva Somayyajula <somayyas@amazon.com>