Chore: Generated Id are now declaration-specific (#5669)
Previously, Id for the compilation of local variables were global, which
caused issues to merge generated files. Fixes #5330

### How has this been tested?
This is a refactoring. Existing tests should pass. Generated code is a
witness that this change performs as expected.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Remy Willems <rwillems@amazon.com>