Chore: Use of concurrentbag instead of list for LSP trees (#5745)
Previously, we used list for the children of a verification tree.
However, processes might concurrently add to this list and try to make a
copy of it, which sometimes triggered issues. By switching to a
concurrentbag, these issues should go away. The list of children is not
ordered anyway. Fixes #5744

### Description
Not visible by users

### How has this been tested?
CI will test that change.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>