fix: Fix various type-system-refresh issues (#5712)
### Description

This PR contains various fixes in the new resolver, found while
preparing Dafny to switch to `--type-system-refresh` as the default.

### How has this been tested?

These issues that this PR fixes were found when running the test suite
with a version of `dafny` that uses `--type-system-refresh
--general-traits=datatype --general-newtypes` by default. However, since
not all test files pass with these options, this PR does not change the
default options, which also means that the existing tests don't test for
the behavior that has been fixed. The purpose of this PR is to get these
fixes in as soon as possible (and to make sure they don't break any
existing functionality). The full testing will happen in a future PR
that changes the default options.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>